% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{Assign UTM Zone}
\alias{Assign UTM Zone}
\alias{assign.utm.zone}
\title{Assign UTM Zones}
\usage{
assign.utm.zone(lat, lon, utm.config)
}
\arguments{
\item{lat,}{lon and utm config dataframe. UTM config for Northern Hemisphere is provided.}
}
\description{
Assigns UTM Zone to given Lat Lon pair
}
\details{
For assigning UTM Zones it does [lat.min, lat.max] & [lon.min, lon.max)
}

