% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{Transform GPS Coords to UTM Coords}
\alias{Transform GPS Coords to UTM Coords}
\alias{transform.gps.to.utm.coords}
\title{Transform GPS Coordinates to UTM Coordinates}
\usage{
\method{transform}{gps.to.utm.coords}(d, utm.config)
}
\arguments{
\item{lat,}{lon and utm config dataframe. UTM config for Northern Hemisphere is provided.}
}
\description{
Transforms GPS Coordinates to UTM Coordinates (East, North)
}

