\name{utms-package}
\alias{utms-package}
\alias{utms}
\docType{package}
\title{
Transform GPS Coordinates to UTM Coordinate System
}
\description{
This packages takes GPS Coordinates, assigns UTM Zone Number 
(currently implemented only for Northern Hemisphere) and then
Transforms GPS Coordinates to UTM Coordinate System.
}
\details{
\tabular{ll}{
Package: \tab utms\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-09-14\cr
License: \tab GNU GPL\cr
}
}
\author{
Siddharth Thakur <sid.thakur@outlook.com>
}
\references{
}

\keyword{ package }
\seealso{

}
\examples{

}
